Added CMakeLists.txt

+++ src/Scanner.h
@@ -30,7 +30,7 @@ private:
 
 private:
 	std::string m_Source;
-	int m_Current, m_Line;
+	size_t m_Current, m_Line;
 	std::vector<Token> m_Tokens;
 	std::unordered_map<std::string, TokenType> m_Keywords = {
 		{"or", TokenType::Or},
